[AI Chat] Introduce ToolFactory and pass instances to ConversationHandler
Creates a Tools factory from the browser to the ConversationHandler via the AIChatService

A minimal stub so that we can parallel some branches of work:
- Tab management Tools
- Content agent Tools
- Basic stateful conversation Tools, e.g. TODO tool, AssistantDetailStorage tool (i.e. created and owned by the ConversationHandler, but have a lifecycle)

\## `ToolProviderFactory`
Introduces `ToolProviderFactory` and `ToolProvider` concept so that different layers can provide Tools and we can control the ownership and lifecycle of them. There are two layers here so that a ToolProvider can maintain state across different tools per-Conversation.

\## `BrowserToolProviderFactory`
The brave/browser layer provides a `BrowserToolProviderFactory` to `AIChatService` which uses it to create a `BrowserToolProvider` for each ConversationHandler.
This is important so that the `BrowserToolProvider` can maintain state across its group of Tool instances and not just per-Tab. For example, the content agent tools are a group of many tools that share state such as which Tabs are being acted upon.

\## Conversation-level Tool instances
This is not used yet but I wanted to provide all the lifecycle options as well as document them since multiple people are working on tools. The use-case here is `components/aichat`-provided but require state and so will be ConversationHandler owned and not global singleton-owned like the existing Tool is via `conversation_tools.h`.